Páginas de cadastro em finalização
Segurança
- Adicionei as validações de senha: letras Maiúsculas e Minúsculas,
  números e símbolos, pelo menos 8 dígitos.
- O arquivo AuthenticatedSessionController.php foi passado para
  diretório app/ e o login está no nosso controle, e não mais apenas no
  do fortify.
- Criei um controller para os temos e políticas.
- Alterei o direcionamento dos arquivos EnsureTermsAccepted.php e
  EnsureUserHasAccessLevel.php para o /calendar em vez de /dashboard,
  que foi removido temporariamente.

Funcionalidades
- Publiquei nos resources os arquivos de email, para possibilitar
  estilização.
- Agora, o aluno recebe senha padrão no cadastro, dentro dos critérios,
  e professores e coordenadores recebem uma senha temporária aleatória
  por email.
- Os emails estão sendo enviados de modo assíncrono por meio de filas no
  banco, tornando o uso e cadastro mais flúido e rápido. Tanto a senah
  temporária quando validação de email e reset de senha seguem esse
  procedimento
- Agora há apenas uma função para iantivar/inativar (toggleStatus), isso
  simplifica e centraliza o método, evitanto redundância.
- No caso dos grupos e bancas, o input de busca para membros ou alunos
  está fazendo uma busca em lista local, todos os alunos ou
  professores/coordenadores são enviados para a view por o método index
  e o js faz a busca nessa lista conforme o usuário digita, isso torna
  tudo mais rápido.
- Os filtros permanecem com fazendo requisição ajax para o back, sem
  busca local.
- Já é possível ocultar ou visualizar senha no ato de login.
- agora o /calendar é uma rota pública, mas seus eventos são
  visualizados apenas aceitando os termos de uso.

Cadastro de bancas
- Adicionei o cadastro de banca, bem como ajustei os relacionamentos
  necessários para isso. por enquanto a rubrica está como nullable, mas
  finalizando elas, será required.
- Alterei algumas relações no banco, committee está reunindo colunas que
  antes eram de professor_committees, que agora será users_committees,
  deivo a necessidade de ter a flexibilidade de cadastrar tanto
  profesores quanto coordenadores na banca, em breve faço a alteração do
  nome, mas a relação já foi alterada.

Geral: banco
- Criei as outras migrations e removi as colunas de alteração para
  datetime2, por que agora usaremos mysql, tudo já está ajustado para
  isso, esqueci de alterar o .env.example, mas substituindo os dados
  relativos a conexão com o banco para mysql já vai rodar.

Estilização
- Os modais em modo update mostram os timestamps do registro, apenas o
  das bancas ainda não está 100% funcional, mas já estou fazendo o
  commit porque faz tempo desde o último.
- Os filtros não são mais selects, estão como dropdown, fica mais
  flexível para estilização.
- Registros inativos aparecem tachados em seus relacionamentos e não são
  exibidos nas listas de create.
- Home Page refeita aos moldes do tailwind + lucide icon para os ícones.

+++ app/Http/Controllers/GroupController.php
@@ -3,6 +3,7 @@
 namespace App\Http\Controllers;
 
 use App\Models\Paper;
+use App\Models\User;
 use Illuminate\Http\Request;
 use Illuminate\Http\JsonResponse;
 use Illuminate\Support\Facades\DB;
@@ -15,52 +16,14 @@ use Random\RandomException;
 use Symfony\Component\HttpFoundation\StreamedResponse;
 
 use Illuminate\Support\Facades\Log;
+use Throwable;
 
 class GroupController extends Controller
 {
     /**
      * @throws RandomException
      */
-    public function index(Request $request): View
-    {
-        //if (!session()->has('custom_token')) {}
-        session(['dynamic_token' => bin2hex(random_bytes(24))]);
-
-        $groups = Group::with([
-            'papers',
-            'students' => function ($query) {
-                $query->whereHas('user', function ($q) {
-                    $q->where('state', 1);
-                })
-                ->with(['user:id,name,state']);
-            }
-        ])->paginate(12);
-
-        // Mapeia os dados manualmente
-        $groupsData = $groups->getCollection()->map(function ($group) {
-            return [
-                'id' => $group->id,
-                'theme' => mb_strtoupper($group->theme),
-                'state' => (int) $group->state,
-                'papers' => $group->papers->map(fn($p) => [
-                    'id' => $p->id,
-                    'title' => $p->title,
-                    'file_path' => $p->file_path,
-                ]),
-                'students' => $group->students->map(fn($s) => [
-                    'ra' => $s->ra,
-                    'name' => mb_strtoupper($s->user->name),
-                ]),
-            ];
-        })->values();
-
-        return view('management.groups', [
-            'groups' => $groupsData,
-            'current_page' => $groups->currentPage(),
-            'last_page' => $groups->lastPage(),
-        ]);
-    }
-
+/*
     public function search(Request $request): JsonResponse
     {
         $q = $request->query('q', '');
@@ -91,7 +54,7 @@ class GroupController extends Controller
             })
         );
     }
-
+*/
     public function showPaper($filename): StreamedResponse
     {
         // Evita acesso fora da pasta
@@ -109,18 +72,72 @@ class GroupController extends Controller
         ]);
     }
 
+    /**
+     * @throws RandomException
+     */
+    public function index(Request $request): View
+    {
+        //if (!session()->has('custom_token')) {}
+        session(['dynamic_token' => bin2hex(random_bytes(24))]);
+
+        $groups = Group::with([
+            'papers',
+            'students.user:id,name,state,updated_at,created_at'
+        ])->orderBy('id')->paginate(12);
+
+        // Mapeia os dados manualmente
+        $groupsData = $groups->getCollection()->map(function ($group) {
+            return [
+                'id' => $group->id,
+                'theme' => mb_strtoupper($group->theme),
+                'state' => (int) $group->state,
+                'papers' => $group->papers->map(fn($p) => [
+                    'id' => $p->id,
+                    'title' => $p->title,
+                    'file_path' => $p->file_path,
+                ]),
+                'students' => $group->students->map(fn($s) => [
+                    'ra' => $s->ra,
+                    'name' => $s->user->name,
+                    'state' => $s->user->state,
+                ]),
+                'created_at' => $group->created_at,
+                'updated_at' => $group->updated_at,
+            ];
+        })->values();
+
+        $students = Student::with(['user:id,name', 'group:id,theme'])
+            ->whereHas('user', function ($sub) {
+                $sub->where('state', 1);
+            })
+            ->select('ra', 'user_id', 'group_id')
+            ->orderBy(
+                User::select('name')
+                    ->whereColumn('users.id', 'students.user_id')
+            )
+            ->get();
+
+        return view('management.groups', [
+            'groups' => $groupsData,
+            'current_page' => $groups->currentPage(),
+            'last_page' => $groups->lastPage(),
+            'students' => $students->map(function ($student) {
+                return [
+                    'ra' => $student->ra,
+                    'name' => $student->user->name ?? '(sem nome)',
+                    'group' => $student->group->theme ?? null,
+                ];
+            })
+        ]);
+    }
+
     public function show(Request $request): JsonResponse
     {
         //DB::enableQueryLog();
         $query = Group::with([
             'papers',
-            'students' => function ($query) {
-                $query->whereHas('user', function ($q) {
-                    $q->where('state', 1);
-                })
-                ->with(['user:id,name,state']);
-            }
-        ]);
+            'students.user:id,name,state,updated_at,created_at'
+        ])->orderBy('id');
 
         if ($request->filled('search')) {
             $search = $request->input('search');
@@ -128,6 +145,9 @@ class GroupController extends Controller
                 $q->where('theme', 'like', "%{$search}%")
                     ->orWhereHas('students.user', function ($sub) use ($search) {
                         $sub->where('name', 'like', "%{$search}%");
+                    })
+                    ->orWhereHas('students', function ($sub) use ($search) {
+                        $sub->where('ra', 'like', "%{$search}%");
                     });
             });
         }
@@ -151,6 +171,16 @@ class GroupController extends Controller
         }
 
         $groups = $query->paginate(12);
+        $students = Student::with(['user:id,name', 'group:id,theme'])
+            ->whereHas('user', function ($sub) {
+                $sub->where('state', 1);
+            })
+            ->select('ra', 'user_id', 'group_id')
+            ->orderBy(
+                User::select('name')
+                    ->whereColumn('users.id', 'students.user_id')
+            )
+            ->get();
         //Log::info('Queries executadas:', DB::getQueryLog());
 
         // Mapeia para retornar somente os campos necessários
@@ -166,8 +196,11 @@ class GroupController extends Controller
                 ]),
                 'students' => $group->students->map(fn($s) => [
                     'ra' => $s->ra,
-                    'name' => mb_strtoupper($s->user->name),
+                    'name' => $s->user->name,
+                    'state' => $s->user->state,
                 ]),
+                'created_at' => $group->created_at,
+                'updated_at' => $group->updated_at,
             ];
         })->values();
 
@@ -175,9 +208,19 @@ class GroupController extends Controller
             'data' => $groupsData,
             'current_page' => $groups->currentPage(),
             'last_page' => $groups->lastPage(),
+            'students' => $students->map(function ($student) {
+                return [
+                    'ra' => $student->ra,
+                    'name' => $student->user->name ?? '(sem nome)',
+                    'group' => $student->group->theme ?? null,
+                ];
+            })
         ]);
     }
 
+    /**
+     * @throws Throwable
+     */
     public function store(Request $request): JsonResponse
     {
         $request->validate([
@@ -193,10 +236,9 @@ class GroupController extends Controller
             ->exists();
 
         if ($alunosEmOutroGrupo) {
-            Log::warning('Aluno(s) já pertence(m) a outro grupo', [
+            /*Log::warning('Aluno(s) já pertence(m) a outro grupo', [
                 'members' => $request->members,
-            ]);
-
+            ]);*/
             return response()->json([
                 'success' => false,
                 'errors' => [
@@ -218,11 +260,6 @@ class GroupController extends Controller
 
             if ($request->hasFile('file')) {
                 $uploadedFile = $request->file('file');
-                Log::debug('Arquivo recebido:', [
-                    'originalName' => $uploadedFile->getClientOriginalName(),
-                    'size' => $uploadedFile->getSize(),
-                    'mimeType' => $uploadedFile->getMimeType(),
-                ]);
                 $filePath = $uploadedFile->store('papers', 'public');
                 $title = pathinfo($uploadedFile->getClientOriginalName(), PATHINFO_FILENAME);
 
@@ -234,7 +271,7 @@ class GroupController extends Controller
             }
         });
 
-        $group->load(['papers', 'students']);
+        $group->load(['papers', 'students.user']);
 
         return response()->json([
             'success' => true,
@@ -250,12 +287,17 @@ class GroupController extends Controller
                 ]),
                 'students' => $group->students->map(fn($s) => [
                     'ra' => $s->ra,
-                    'name' => mb_strtoupper($s->user->name),
-                ])
+                    'name' => $s->user->name,
+                ]),
+                'created_at' => $group->created_at,
+                'updated_at' =>  $group->updated_at,
             ]
         ]);
     }
 
+    /**
+     * @throws Throwable
+     */
     public function update(Request $request, $id): JsonResponse
     {
         if(!$id) {
@@ -266,7 +308,7 @@ class GroupController extends Controller
 
         $request->validate([
             'theme' => "required|string|max:255|unique:groups,theme,{$id},id",
-            'file' => 'nullable|file|mimes:pdf|max:5120',
+            'file' => 'nullable|file|mimes:pdf|max:10240',
             'members' => 'required|array|min:1',
             'members.*' => 'required|string|exists:students,ra',
         ]);
@@ -298,46 +340,74 @@ class GroupController extends Controller
 
         DB::transaction(function () use ($request, $group) {
             // Atualiza o tema
-            $group->theme = $request->theme;
-            $group->save();
+            if($group->theme != $request->theme) {
+                $group->update([
+                    'theme' => $request->theme,
+                ]);
+            }
 
             // Remove o group_id dos alunos que não estão mais no grupo
-            Student::where('group_id', $group->id)
+            $removed = Student::where('group_id', $group->id)
                 ->whereNotIn('ra', $request->members)
                 ->update(['group_id' => null]);
 
+            if($removed > 0) $group->touch();
+
             // Atualiza o group_id dos alunos selecionados
-            Student::whereIn('ra', $request->members)
+            $added = Student::whereIn('ra', $request->members)
+                ->where(function ($q) use ($group) {
+                    $q->whereNull('group_id')
+                        ->orWhere('group_id', '<>', $group->id);
+                })
                 ->update(['group_id' => $group->id]);
+
+            if($added > 0) $group->touch();
 
             // Se tem arquivo novo, salva e cria novo Paper, removendo antigo
             if ($request->hasFile('file')) {
                 $uploadedFile = $request->file('file');
-                /*Log::debug('Arquivo recebido:', [
-                    'originalName' => $uploadedFile->getClientOriginalName(),
-                    'size' => $uploadedFile->getSize(),
-                    'mimeType' => $uploadedFile->getMimeType(),
-                ]);*/
-                $filePath = $uploadedFile->store('papers', 'public');
-                $title = pathinfo($uploadedFile->getClientOriginalName(), PATHINFO_FILENAME);
 
-                // Remove papers antigos e apaga o arquivo também)
-                $oldPapers =  $group->papers;
-                foreach ($oldPapers as $oldPaper) {
-                    Storage::disk('public')->delete($oldPaper->file_path);
+                $oldPaper = $group->papers()->latest()->first();
+
+                // Hash real do novo arquivo (SHA1 do conteúdo)
+                $newHash = sha1_file($uploadedFile->getRealPath());
+
+                if ($oldPaper) {
+                    // Lê o arquivo atual no storage e calcula hash
+                    $oldHash = sha1(Storage::disk('public')->get($oldPaper->file_path));
+
+                    if ($newHash === $oldHash) {
+                        // Mesmo arquivo → só atualiza título se mudou
+                        $updated = $oldPaper->update([
+                            'title' => pathinfo($uploadedFile->getClientOriginalName(), PATHINFO_FILENAME),
+                        ]);
+
+                        return; // não mexe no banco nem cria arquivo novo
+                    }
+
+                    // Arquivo diferente → remove os antigos
+                    foreach ($group->papers as $paper) {
+                        Storage::disk('public')->delete($paper->file_path);
+                    }
+                    $group->papers()->delete();
+
+                    $group->touch();
                 }
 
-                $group->papers()->delete();
+                // Só agora salva porque sabemos que é realmente outro arquivo
+                $newFilePath = $uploadedFile->store('papers', 'public');
 
                 Paper::create([
-                    'title' => $title,
-                    'file_path' => $filePath,
+                    'title' => pathinfo($uploadedFile->getClientOriginalName(), PATHINFO_FILENAME),
+                    'file_path' => $newFilePath,
                     'group_id' => $group->id,
                 ]);
             }
+
         });
 
-        $group->load(['papers', 'students.user']);
+        // Recarrega completamente grupo + alunos + papers
+        $group = Group::with(['papers', 'students.user'])->find($id);
 
         return response()->json([
             'success' => true,
@@ -355,11 +425,13 @@ class GroupController extends Controller
                     'ra' => $s->ra,
                     'name' => mb_strtoupper($s->user->name),
                 ]),
+                'created_at' => $group->created_at,
+                'updated_at' => $group->updated_at,
             ]
         ]);
     }
 
-    public function inactivate($id): JsonResponse
+    public function toggleStatus($id,$action): JsonResponse
     {
         $group = Group::find($id);
 
@@ -369,33 +441,24 @@ class GroupController extends Controller
             ], 422);
         }
 
-        $group->update(['state' => 0]);
-
-        $group->refresh();
-
-        return response()->json([
-            'success' => true,
-            'message' => 'Grupo inativado com sucesso!'
-        ]);
-    }
-
-    public function activate($id): JsonResponse
-    {
-        $group = Group::find($id);
-
-        if (!$group) {
+        if ($action === 'inactivate') {
+            $group->update(['state' => 0]);
+        } elseif ($action === 'activate') {
+            $group->update(['state' => 1]);
+        }  else {
             return response()->json([
-                'message' => 'Grupo não encontrado!'
+                'message' => 'Ação inválida!'
             ], 422);
         }
 
-        $group->update(['state' => 1]);
-
         $group->refresh();
 
         return response()->json([
             'success' => true,
-            'message' => 'Grupo ativado com sucesso!'
+            'message' => 'Grupo atualizado com sucesso!',
+            'state' => (int) $group->state,
+            'created_at' => $group->created_at,
+            'updated_at' => $group->updated_at,
         ]);
     }
 }
